libc: add basical localtime support
We now pass the right things around to make this work.

At the moment the only format understood is zone+-shift in hours. So for example
TZ=BST+1.

+++ Library/include/time.h
@@ -40,7 +40,7 @@ extern time_t mktime(struct tm * __tp);
 extern double difftime(time_t *__time2, time_t *__time1);
 
 
-extern void __tm_conv(struct tm *tmbuf, time_t *t, int offset);
+extern void __tm_conv(struct tm *tmbuf, time_t *t, long offset);
 extern char *asctime(struct tm * __tp);
 extern char *asctime_r(struct tm *, char * __buf);
 extern char *ctime(time_t * __tp);
